requirements.txt
pyyaml
requests
irc
openai
cryptography
